we can now split atomWithReset and useResetAtom
* it wasn't possible before since RESET wasn't exported
